Recognize periods after numbers that are not followed by numbers as periods
So, instead of '4.' being recognized as a tokenization failure, we emit
an INTEGER token of '4' and a PERIOD.

This also applies to binary/octal/hexadecimal floats. So '0x.' will be
an ERROR token of '0x' and a PERIOD. The reason that the former is an
error is because b/o/x numbers have to have a number in them instead of
just being the format specifier.

And as a benefit, the code does not look like hog shit.